refactor: SignUpForm 컴포넌트를 제거하고 SignUpPage로 통합하여 코드 간결화 및 UI 개선
